230808 1. 충돌 버그 수정 (세영이 감사) 2. 카메라 버그 수정 (이지섭님 감사) - 카메라가 플레이어를 따라다니는 기능 다시 추가 - 씬 진입 후 카메라가 플레이어를 향해 움직이는 문제 수정 (SetTarget하면 즉시 타겟을 바라보도록 수정) -> 월드를 벗어나는 것 체크는 안했음. 동작에 문제 없기 때문에.. 추후 문제가 발생하면 추가 3. 보스 몬스터 클래스만 추가함

+++ Engine_SOURCE/Lu_PaintShader.cpp
@@ -16,8 +16,8 @@ namespace Lu::graphics
 	{
 		m_Target->BindUnorderedAccessViews(0);
 		 
-		m_GroupX = m_Target->GetWidth() / m_ThreadGroupCountX + 1;
-		m_GroupY = m_Target->GetHeight() / m_ThreadGroupCountY + 1;
+		m_GroupX = (UINT)m_Target->GetWidth() / m_ThreadGroupCountX + 1;
+		m_GroupY = (UINT)m_Target->GetHeight() / m_ThreadGroupCountY + 1;
 		m_GroupZ = 1;
 	}
